+++ app/src/androidTest/java/com/example/mobillaborapp/test/AddPictureActivityTest.kt
@@ -1,4 +1,4 @@
-package com.example.mobillaborapp.ui.picturedetails
+package com.example.mobillaborapp.test
 
 import android.content.Intent
 import androidx.test.espresso.Espresso
@@ -6,6 +6,7 @@ import androidx.test.espresso.assertion.ViewAssertions
 import androidx.test.espresso.matcher.ViewMatchers
 import androidx.test.ext.junit.runners.AndroidJUnit4
 import com.example.mobillaborapp.R
+import com.example.mobillaborapp.ui.addpicture.AddPictureActivity
 import com.example.mobillaborapp.utils.AndroidTestUtils
 import com.example.mobillaborapp.utils.EspressoTest
 import org.junit.After
@@ -15,27 +16,21 @@ import org.junit.Test
 import org.junit.runner.RunWith
 
 @RunWith(AndroidJUnit4::class)
-class PictureDetailsActivityTest : EspressoTest<PictureDetailsActivity>(PictureDetailsActivity::class.java) {
+class AddPictureActivityTest : EspressoTest<AddPictureActivity>(AddPictureActivity::class.java){
 
     @Before
     fun setUp() {
         AndroidTestUtils.setTestInjector()
-        activityRule.launchActivity(Intent().putExtra("IMAGE_ID", PictureDetailsActivityTest.IMAGEID))
+        activityRule.launchActivity(Intent())
     }
 
     @Test
-    fun testSearch() {
-        Espresso.onView(ViewMatchers.withId(R.id.tvBreed))
-            .check(ViewAssertions.matches(ViewMatchers.withText(PictureDetailsActivityTest.BreedName)))
-        /*
-        waitFor(2000) {
-            onView(allOf(withId(com.example.mobillaborapp.R.id.tvName), withText(ARTIST))).check(matches(isDisplayed()))
-        }
-         */
+    fun testBreedSpinner() {
+        Espresso.onView(ViewMatchers.withId(R.id.mySpinner))
+                .check(ViewAssertions.matches(ViewMatchers.withSpinnerText(AddPictureActivityTest.BreedName)))
     }
 
     companion object {
-        private const val IMAGEID = "2u0QIn3hP"
         private const val BreedName = "Bambino"
     }
